+++ ppsp.py
@@ -1,4 +1,4 @@
-import os, sys, queue, re, signal, subprocess, threading, time
+import os, sys, queue, re, subprocess, threading
 
 
 class PPSP:
@@ -11,9 +11,10 @@ class PPSP:
 
     class _Status:
         running = False
+        stopping = False
         stopped = True
 
-    def __init__(self, shell_command: str, exit_condition: str = None):
+    def __init__(self, shell_command: str, verbose: bool = False, exit_condition: str = None):
         # Subprocess management
         self._shell_command = shell_command
 
@@ -21,7 +22,9 @@ class PPSP:
             self._exit_condition = re.compile(exit_condition) \
                 if exit_condition is not None else None
         except re.error as e:
-            print('Regular expression did not compile. Skipping...')
+            if verbose:
+                print('Regular expression did not compile. Skipping...')
+            
             self._exit_condition = None
         
         self._exit_condition_met = False
@@ -49,18 +52,19 @@ class PPSP:
         self.__start_process()
         self._status.running = True
         self._status.stopped = False
+        self._status.stopping = False
 
         self._status_tr.start()
         self._reader_tr.start()
         self._writer_tr.start()
 
     def stop(self) -> None:
-        self._status.running = False
-
         self._status_tr.join()
         self._reader_tr.join()
         self._writer_tr.join()
 
+        self._status.running = False
+        self._status.stopping = False
         self._status.stopped = True
 
         if self._exit_condition_met:
@@ -69,6 +73,9 @@ class PPSP:
     def is_running(self) -> bool:
         return self._status.running
     
+    def is_stopping(self) -> bool:
+        return self._status.stopping
+
     def is_stopped(self) -> bool:
         return self._status.stopped
 
@@ -85,81 +92,95 @@ class PPSP:
         while True:
             if not self._subprocess:
                 continue
-
-            if self._subprocess.poll() is not None:
-                self._status.stopped = True
+            
+            if self._status.stopping:
                 break
 
+            if self._subprocess.poll() is not None:
+                self._status.stopping = True
+            
     def __reader(self) -> None:
         while True:
             if not self._subprocess:
                 continue
 
-            if self._status.stopped:
+            if self._status.stopping:
                 break
             
             if self._subprocess.stdout:
                 for linen in self._subprocess.stdout:
+                    if self._status.stopping:
+                        break
                     line = linen.decode(sys.getdefaultencoding())
                     if line:
                         print(line, end='')
                         if self._exit_condition:
                             if re.match(self._exit_condition, line):
                                 self._exit_condition_met = True
-                                self._status.stopped = True
-                                break
+                                self._status.stopping = True
 
     def __writer(self) -> None:
         while True:
             if not self._subprocess:
                 continue
 
-            if self._status.stopped:
+            if self._status.stopping:
                 break
 
             if not self._queue.empty():
                 self.__send_input(self._queue.get())
 
 
-def ppsp(shell_command: str, exit_condition: str = None) -> None:
+def ppsp(shell_command: str, verbose: bool = False, exit_condition: str = None) -> None:
+    if verbose:
+        print('PPSP starting... Press CTRL+C to quit')
+
     ppsp_instance = None
     try:
-        ppsp_instance = PPSP(shell_command, exit_condition)
+        ppsp_instance = PPSP(shell_command, verbose, exit_condition)
         ppsp_instance.start()
         ppsp_instance.stop()
     except KeyboardInterrupt:
-        print('Keyboard Interrupt. Exiting PPSP gracefully...')
+        if verbose:
+            print('Keyboard Interrupt. Exiting PPSP gracefully...')
+
         if ppsp_instance is not None:
             if ppsp_instance.subprocess is not None:
                 try:
-                    os.killpg(os.getpgid(ppsp_instance.subprocess.pid), signal.SIGTERM)
+                    ppsp_instance.subprocess.kill()
                 except (OSError, ProcessLookupError):
                     pass
         sys.exit(0)
     except PPSP.Exceptions.PPSPExitException:
-        print('Exit condition met. Exiting PPSP gracefully...')
+        if verbose:
+            print('Exit condition met. Exiting PPSP gracefully...')
+
         if ppsp_instance is not None:
             if ppsp_instance.subprocess is not None:
                 try:
-                    os.killpg(os.getpgid(ppsp_instance.subprocess.pid), signal.SIGTERM)
+                    ppsp_instance.subprocess.kill()
                 except (OSError, ProcessLookupError):
                     pass
         sys.exit(0)
     except Exception:
-        print('Unhandled exception occured. Exiting PPSP gracefully...')
+        if verbose:
+            print('Unhandled exception occured. Exiting PPSP gracefully...')
+        
         if ppsp_instance is not None:
             if ppsp_instance.subprocess is not None:
                 try:
-                    os.killpg(os.getpgid(ppsp_instance.subprocess.pid), signal.SIGTERM)
+                    ppsp_instance.subprocess.kill()
                 except (OSError, ProcessLookupError):
                     pass
         sys.exit(0)
     else:
-        print('Finished normally. Exiting PPSP gracefully...')
+        if verbose:
+            print('Finished normally. Exiting PPSP gracefully...')
+        
         if ppsp_instance is not None:
             if ppsp_instance.subprocess is not None:
                 try:
-                    os.killpg(os.getpgid(ppsp_instance.subprocess.pid), signal.SIGTERM)
+                    ppsp_instance.subprocess.kill()
                 except (OSError, ProcessLookupError):
                     pass
         sys.exit(0)