+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 Changes between project versions will be documented here.
 
 
+## [1.0.1]
+### Added
+- Stopping status flag
+
+### Changed 
+- Command line arguments for example.py
+
+### Fixed
+- PPSP would not exit after subprocess was finished executing
+
+
 ## [1.0.0]
 ### Added 
 - PPSP class for running a persistent shell, allowing for stdout or stdin